@@ -9,7 +9,7 @@
   },
   "dependencies": {
     "core-js": "^3.8.3",
-    "vue": "^2.6.14"
+    "vue": "^3.2.36"
   },
   "devDependencies": {
     "@babel/core": "^7.12.16",
